fix(build): extract treetime-schema crate to fix cross-compilation
- app-cli build script depended on treetime for schema generation, which transitively pulls in openblas-src. Build scripts run on the host but cross-compilation containers only have target-architecture openblas, causing linker failures on aarch64-linux-gnu, aarch64-linux-musl, and x86_64-windows-gnu
- Extract ProgressEvent, ErrorResponse, VersionInfo, and generate_schema into a new treetime-schema crate with no openblas dependency. Consumers import from treetime_schema directly (no re-exports through treetime)
