+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
   "packages/treetime-grid",
   "packages/treetime-io",
   "packages/treetime-ops",
+  "packages/treetime-schema",
   "packages/treetime-primitives",
   "packages/treetime-utils",
   "packages/treetime-validation",
@@ -153,6 +154,7 @@ treetime-grid = { path = "packages/treetime-grid" }
 treetime-io = { path = "packages/treetime-io" }
 treetime-ops = { path = "packages/treetime-ops" }
 treetime-primitives = { path = "packages/treetime-primitives" }
+treetime-schema = { path = "packages/treetime-schema" }
 treetime-utils = { path = "packages/treetime-utils" }
 util-augur-node-data-json = { path = "packages/util-augur-node-data-json" }
 util-phyloxml = { path = "packages/util-phyloxml" }
