+++ packages/app-server/Cargo.toml
@@ -13,6 +13,7 @@ version.workspace = true
 [dependencies]
 app-api = { workspace = true }
 treetime = { workspace = true }
+treetime-schema = { workspace = true }
 treetime-utils = { workspace = true }
 
 async-stream = { workspace = true }
