+++ packages/app-server/src/routes.rs
@@ -5,7 +5,7 @@ use crate::error::AppError;
 use crate::sse::handle_command;
 use crate::state::ServerConfig;
 use app_api::datasets::discover_datasets;
-use app_api::version::version_info;
+use treetime_schema::version_info;
 use axum::extract::State;
 use axum::response::Response;
 use axum::routing::{get, post};
