+++ packages/treetime-schema/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "treetime-schema"
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+publish.workspace = true
+repository.workspace = true
+version.workspace = true
+
+
+[dependencies]
+treetime-utils = { workspace = true }
+
+clap = { workspace = true, optional = true }
+eyre = { workspace = true }
+log = { workspace = true }
+schemars = { workspace = true }
+serde = { workspace = true }
+strum = { workspace = true }
+strum_macros = { workspace = true }
+
+[features]
+clap = ["dep:clap"]
+
+[lints]
+workspace = true
