+++ packages/treetime-schema/src/lib.rs
@@ -1,0 +1,7 @@
+mod progress;
+mod schema;
+mod version;
+
+pub use progress::{ErrorResponse, ProgressEvent};
+pub use schema::{TreetimeSchemaFormat, generate_schema};
+pub use version::{VersionInfo, version_info};
