+++ packages/treetime-schema/src/progress.rs
@@ -1,0 +1,15 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct ProgressEvent {
+  pub stage: String,
+  pub fraction: f64,
+  pub message: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct ErrorResponse {
+  pub code: String,
+  pub message: String,
+}
